refactor(test): split SignalGraph tests out of test_host.cpp (P11-5, #2647)
test_host.cpp was 2,190 lines — over the long-tail 2,000-line target. The
SignalGraph cluster (35 of 56 cases, ~1,335 lines) is the dominant concern;
carve it verbatim into test_host_signal_graph.cpp (own executable, no CLAP
fixture needed — pure graph routing/topology, not plugin loading). The
interleaved per-test helper namespaces travel with the block.

Parent test_host.cpp: 2,190 -> 853 lines (21 cases: scanner/slot/serializer/
param-queue, incl. the CLAP-fixture-guarded ClapSlot tests). Added the
graph_serializer.hpp include to the parent (it lived inside the extracted
range). Coverage preserved: 17 (parent, fixture off) + 35 (SignalGraph) = 52
run locally GPU-off; the 4 #ifdef PULP_TEST_CLAP_PATH ClapSlot cases stay in
the parent and run with the fixture in CI → 56 total, unchanged.
