+++ test/CMakeLists.txt
@@ -2880,6 +2880,13 @@ if(TARGET PulpGain_CLAP)
 endif()
 catch_discover_tests(pulp-test-host)
 
+# P11-5 (#2647): SignalGraph tests carved out of test_host.cpp to bring the
+# 2,190-line parent under 2,000 (now 852). No CLAP fixture needed — these
+# exercise pure graph routing/topology, not plugin loading.
+add_executable(pulp-test-host-signal-graph test_host_signal_graph.cpp)
+target_link_libraries(pulp-test-host-signal-graph PRIVATE pulp::host Catch2::Catch2WithMain)
+catch_discover_tests(pulp-test-host-signal-graph)
+
 # Host regression coverage (#52) — end-to-end scan / load / process /
 # unload lifecycle, scanner failure modes (moduleinfo.json + manifest.ttl),
 # state round-trip, and hot-reload race-cleanliness. Separate binary so a
